refactor: Move NetworkMonitor to core:network module
